Move `parse_instruction()` to InstructionTemplate.
