+++ Instruction.py
@@ -3,15 +3,9 @@
 # SPDX-License-Identifier: LGPL-3.0-only
 
 import re
-from copy import deepcopy
 
-import PluginInfo
-from Immediate import Immediate
-from ImplementationException import ImplementationException
 from InstructionTemplate import InstructionTemplate, LoopMembership
-from Operand import Operand, OperandType, SparseMask
 from InstructionEncoding import InstructionEncoding
-from Register import Register
 from helperFunctions import normalize_llvm_syntax, list_to_int
 
 
@@ -92,95 +86,6 @@ class Instruction(InstructionTemplate):
         self.loop_member = self.get_loop_membership(self.llvm_syntax)
 
         self.parse_instruction()
-
-    def parse_instruction(self) -> None:
-        """Parses all operands of the instruction which are encoded."""
-
-        self.llvm_filtered_operands = self.remove_invisible_in_out_regs(
-            self.llvm_syntax, deepcopy(self.llvm_in_out_operands)
-        )
-        self.operand_indices = self.get_syntax_operand_indices(self.llvm_syntax, self.llvm_filtered_operands)
-
-        # Update syntax indices
-        if self.has_new_non_predicate:
-            op_name = self.llvm_in_out_operands[self.new_operand_index][1]
-            self.new_operand_index = self.operand_indices[op_name]
-            # log("{}\nnew: {}".format(self.llvm_syntax, self.new_operand_index), LogLevel.DEBUG)
-        if self.has_extendable_imm:
-            op_name = self.llvm_in_out_operands[self.ext_operand_index][1]
-            self.ext_operand_index = self.operand_indices[op_name]
-            # log("{}\next: {}".format(self.llvm_syntax, self.ext_operand_index), LogLevel.DEBUG)
-
-        if len(self.llvm_filtered_operands) > PluginInfo.MAX_OPERANDS:
-            warning = "{} instruction struct can only hold {} operands. This" " instruction has {} operands.".format(
-                PluginInfo.FRAMEWORK_NAME,
-                PluginInfo.MAX_OPERANDS,
-                len(self.llvm_filtered_operands),
-            )
-            raise ImplementationException(warning)
-
-        # TODO Some instructions encode some register explicitly in the syntax. At the moment we do not,
-        #  but maybe should add them here somehow as registers. Example: J4_cmpeq_fp0_jump_t
-        #  But note that they don't seem to have an index attached to them.
-
-        # TODO Parse high/low access of registers.
-
-        for in_out_operand in self.llvm_filtered_operands:
-            op_name = in_out_operand[1]
-            op_type = in_out_operand[0]["def"]
-            syntax_index = self.operand_indices[op_name]
-
-            # Parse register operand
-            if Operand.get_operand_type(op_type) is OperandType.REGISTER:
-                # Indices of new values (stored in "opNewValue") are only for non predicates.
-                is_new_value = self.new_operand_index == syntax_index and self.has_new_non_predicate
-                operand = Register(op_name, op_type, is_new_value, syntax_index)
-                # Whether the predicate registers holds a new value is denoted in "isPredicatedNew".
-                if self.is_pred_new and operand.is_predicate:
-                    operand.is_new_value = True
-            # Parse immediate operands
-            elif Operand.get_operand_type(op_type) is OperandType.IMMEDIATE:
-                extendable = self.has_extendable_imm and self.ext_operand_index == syntax_index
-                operand = Immediate(
-                    op_name,
-                    op_type,
-                    extendable,
-                    self.extendable_alignment,
-                    syntax_index,
-                )
-
-            else:
-                raise ImplementationException("Unknown operand type: {}, op_name: {}".format(op_type, op_name))
-
-            if op_name in self.constraints:
-                operand.is_in_out_operand = True
-                operand.is_out_operand = True
-                operand.is_in_operand = True
-            elif in_out_operand in self.llvm_in_operands:
-                operand.is_in_operand = True
-            elif in_out_operand in self.llvm_out_operands:
-                operand.is_out_operand = True
-
-            # Add opcode extraction code
-            if operand.type == OperandType.IMMEDIATE and operand.is_constant:  # Constants have no parsing code.
-                pass
-            else:
-                if operand.is_in_out_operand and op_name[-2:] == "in":  # In/Out Register
-                    mask = self.encoding.operand_masks[op_name[:-2]]  # Ends with "in"
-                else:
-                    mask = self.encoding.operand_masks[op_name]
-                operand.opcode_mask = SparseMask(mask)
-
-            # On the fly check whether the new values have been assigned correctly.
-            if op_name + ".new" in self.llvm_syntax:
-                if not operand.is_new_value:
-                    raise ImplementationException(
-                        "Register has new value in syntax but not as object."
-                        + "It has been parsed incorrectly! Are the indices"
-                        " correctly set?" + "Affected instruction: {}".format(self.llvm_syntax)
-                    )
-
-            self.operands[op_name] = operand
 
     @staticmethod
     def get_num_operands(llvm_syntax: str, llvm_operands: list) -> int:
